# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Funding Application Platform API
  version: 1.0.0
  description: |
    RESTful API for the Funding Application Submission & Assessment Platform.

    This API enables:
    - **Applicants** to submit applications with uploaded files and confirmations
    - **Coordinators** to manage funding calls, assign assessors, and view master results
    - **Assessors** to score assigned applications against defined criteria

    ## Authentication

    All endpoints except public call listings require JWT Bearer authentication.
    Service-to-service communication uses API key authentication.

    ## Rate Limiting

    - Standard endpoints: 100 requests/minute per user
    - File upload endpoints: 10 requests/minute per user
    - Export endpoints: 5 requests/minute per user

    ## Timezone Handling

    All timestamps are stored in UTC and returned in ISO 8601 format.
    Deadlines are enforced in Europe/London timezone.
  contact:
    name: Platform Support
    email: support@fundingplatform.example.com
    url: https://fundingplatform.example.com/support
  license:
    name: Proprietary
    url: https://fundingplatform.example.com/license
  x-logo:
    url: https://fundingplatform.example.com/logo.png
    altText: Funding Platform Logo
servers:
  - url: https://api.fundingplatform.example.com/api/v1
    description: Production server
  - url: https://staging-api.fundingplatform.example.com/api/v1
    description: Staging server
...
security:
  - BearerAuth: []
paths:
...
  /assignments/bulk:
    post:
      tags:
        - Assignments
      summary: Bulk assign applications
      description: |
        Assigns multiple applications to assessors using the specified strategy.

        **Required role:** Coordinator
      operationId: bulkAssign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - call_id
                - strategy
              properties:
                call_id:
                  type: string
                  format: uuid
                strategy:
                  type: string
                  enum: [round_robin, balanced, manual]
                  description: |
                    - `round_robin`: Distribute evenly across all assessors
                    - `balanced`: Consider current workload
                    - `manual`: Use explicit assignments array
                assessor_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                  description: Assessors to include (if not specified, uses entire pool)
                application_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                  description: Applications to assign (if not specified, assigns all unassigned)
                assignments:
                  type: array
                  items:
                    type: object
                    properties:
                      application_id:
                        type: string
                        format: uuid
                      assessor_id:
                        type: string
                        format: uuid
                  description: Explicit assignments (for manual strategy)
                assessors_per_application:
                  type: integer
                  minimum: 1
                  maximum: 10
                  default: 2
                notify_assessors:
                  type: boolean
                  default: true
            example:
              call_id: "550e8400-e29b-41d4-a716-446655440000"
              strategy: "round_robin"
              assessors_per_application: 2
              notify_assessors: true
      responses:
        '200':
          description: Bulk assignment results
          content:
            application/json:
              schema:
                type: object
                properties:
                  created:
                    type: integer
                    description: Number of assignments created
                  skipped:
                    type: integer
                    description: Number of assignments skipped (already exists)
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        application_id:
                          type: string
                        reason:
                          type: string
                  assignments:
                    type: array
                    items:
                      $ref: '#/components/schemas/Assignment'
              example:
                created: 90
                skipped: 0
                errors: []
                assignments:
                  - id: "assign-001"
                    application_id: "app-001"
                    assessor_id: "assessor-001"
                    status: "pending"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service communication
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "Request validation failed"
              details:
                fields:
                  - field: "email"
                    message: "Invalid email format"
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Authentication required"
              details: null
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "FORBIDDEN"
              message: "You do not have permission to perform this action"
              details:
                required_role: "coordinator"
...
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Human-readable error message
              example: "Request validation failed"
            details:
              type: object
              nullable: true
              description: Additional error details
              additionalProperties: true

    # --------------------------------------------------------------------------
    # PAGINATION SCHEMA
    # --------------------------------------------------------------------------
...
    Assessment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        assignment_id:
          type: string
          format: uuid
        application_id:
          type: string
          format: uuid
        application_reference:
          type: string
        assessor_id:
          type: string
          format: uuid
          description: Only visible to coordinators
        status:
          type: string
          enum: [draft, submitted, returned]
        scores:
          type: array
          items:
            $ref: '#/components/schemas/Score'
        total_score:
          type: number
          description: Sum of all criterion scores
        weighted_score:
          type: number
          description: Weighted total score
        overall_comment:
          type: string
        coi_declared:
          type: boolean
        coi_has_conflict:
          type: boolean
        coi_details:
          type: string
        submitted_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Score:
      type: object
      properties:
        criterion_id:
          type: string
        criterion_name:
          type: string
        score:
          type: integer
        max_points:
          type: integer
        weight:
          type: number
        comment:
          type: string
...
    Assignment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        application_id:
          type: string
          format: uuid
        application_reference:
          type: string
        assessor_id:
          type: string
          format: uuid
        assessor_name:
          type: string
          description: Only visible to coordinators
        status:
          type: string
          enum: [pending, in_progress, completed]
        due_at:
          type: string
          format: date-time
        assigned_at:
          type: string
          format: date-time
        assessment:
          $ref: '#/components/schemas/Assessment'
...